Revert to not use NetworkInstance
