+++ backend/Controller.cs
@@ -4,6 +4,8 @@ using static Microsoft.AspNetCore.Http.StatusCodes;
 using backend.models;
 using System.Linq;
 using backend.drawables;
+using System.Collections.Concurrent;
+using System.Threading.Tasks;
 
 namespace backend
 {
@@ -13,29 +15,23 @@ namespace backend
     public class Controller : ControllerBase
     {
 
+        private static ConcurrentBag<Link> links = new ConcurrentBag<Link>();
+        private static ConcurrentBag<Node> processedServers = new ConcurrentBag<Node>();
+        private static HashSet<string> missingServerIds = new HashSet<string>();
+
+
         [HttpGet("/nodes")]
         [ProducesResponseType(Status200OK)]
         public ActionResult<IEnumerable<Node>> GetNodes()
         {   
-            return Program.instances[0].processedServers;
-        }
-
-        [HttpGet("/links")]
-        [ProducesResponseType(Status200OK)]
-        public ActionResult<IEnumerable<Link>> GetLinks()
-        {   
-            return Program.instances[0].links;
+            return processedServers;
         }
 
         [HttpGet("/varz")]
         [ProducesResponseType(Status200OK)]
         public ActionResult<IEnumerable<Server>> GetVarz()
         {   
-            var instance = (from entry in Program.instances
-                where entry.id == 0
-                select entry).FirstOrDefault();
-
-            return instance.servers;
+            return Program.servers;
         }
 
         // TODO use to get info about server, when it is clicked in UI
@@ -43,9 +39,7 @@ namespace backend
         [ProducesResponseType(Status200OK)]
         public ActionResult<Server> GetVarz([FromRoute] string serverId)
         {   
-            var instance = Program.instances[0];
-
-            var query = from server in instance.servers.AsParallel()
+            var query = from server in Program.servers.AsParallel()
                 where server.server_id == serverId
                 select server;
             return query.FirstOrDefault();
@@ -55,28 +49,76 @@ namespace backend
         [ProducesResponseType(Status200OK)]
         public ActionResult<IEnumerable<Connection>> GetConnz()
         {   
-            return Program.instances[0].connections;
+            return Program.connections;
+        }
+        
+        [HttpGet("/links")]
+        [ProducesResponseType(Status200OK)]
+        public ActionResult<IEnumerable<Link>> GetLinks()
+        {   
+            return links;
         }
 
         [HttpGet("/routez")]
         [ProducesResponseType(Status200OK)]
         public ActionResult<IEnumerable<Route>> GetRoutez()
         {   
-            return Program.instances[0].routes;
+            return Program.routes;
         }
         
         [HttpGet("/gatewayz")]
         [ProducesResponseType(Status200OK)]
         public ActionResult<IEnumerable<Gateway>> GetGatewayz()
         {   
-            return Program.instances[0].gateways;
+            return Program.gateways;
         }
 
         [HttpGet("/leafz")]
         [ProducesResponseType(Status200OK)]
         public ActionResult<IEnumerable<Leaf>> GetLeafz()
         {   
-            return Program.instances[0].leafs;
+            return Program.leafs;
         }
+
+        public static void ProcessData() 
+        {
+            Parallel.ForEach(Program.servers, server => {
+                processedServers.Add(new Node {
+                    server_id = server.server_id, 
+                    server_name = server.server_name, 
+                    ntv_error = false 
+                });
+            });
+
+            // Information about routes are also on server, no request to routez necessary
+            // Maybe info on routes is more up-to-date?
+            Parallel.ForEach(Program.servers, server => {
+                var source = server.server_id;
+                Parallel.ForEach(server.route.routes, route => {
+                    var target = route.remote_id;
+                    links.Add(new Link(source, target));
+                });
+            });
+
+            foreach (var link in links) 
+            {
+                if (!Program.idToServer.ContainsKey(link.target))
+                {
+                    missingServerIds.Add(link.target);
+                    link.ntv_error = true;
+                    links.Add(link);
+                }
+            }
+
+            // Patch for a missing node from varz
+            // TODO dynamically handle these types of errors
+            foreach(var serverId in missingServerIds)
+            {
+                processedServers.Add(new Node{server_id = serverId, ntv_error = true});
+            }
+
+
+        }
+
     }
 }
